add crud menu catatan siswa

+++ src/views/admin/sekolah/submenu/catatanpengembangandiri/Index.vue
@@ -12,7 +12,6 @@ const router = useRouter();
 const route = useRoute();
 
 const id = route.params.id;
-const id2 = route.params.id2;
 
 const dataAsli = ref([]);
 const dataKelas = ref([]);
@@ -51,7 +50,7 @@ getDataKelas();
 const getData = async () => {
   try {
     const response = await Api.get(
-      `admin/datasekolah/${id}/siswa/${id2}/catatankasussiswa`
+      `admin/datasekolah/${id}/catatanpengembangandiri/siswa`
     );
     // console.log(response);
     dataAsli.value = response.data;
@@ -88,35 +87,33 @@ const columns = [
     thClass: "text-center",
   },
   {
+    label: "Nama Siswa",
+    field: "nama",
+    type: "String",
+  },
+  {
+    label: "Kelas",
+    field: "kelas",
+    type: "String",
+  },
+  {
     label: "Kasus",
-    field: "kasus",
-    type: "String",
-  },
-  {
-    label: "Pengambilan Data",
-    field: "pengambilandata",
-    type: "String",
-  },
-  {
-    label: "Teknik Konseling",
-    field: "teknikkonseling",
+    field: "jml",
     type: "String",
   },
 ];
 
-const doEditData = async (id3, index) => {
+const doEditData = async (id2, index) => {
   router.push({
-    name: "AdminSekolahDetailCatatanKasusSiswaEdit",
-    params: { id, id2, id3 },
+    name: "AdminSekolahDetailSiswaEdit",
+    params: { id: id, id2: id2 },
   });
 };
 
-const doDeleteData = async (id3, index) => {
+const doDeleteData = async (id2, index) => {
   if (confirm("Apakah anda yakin menghapus data ini?")) {
     try {
-      const response = await Api.delete(
-        `admin/datasekolah/${id}/siswa/${id2}/catatankasussiswa/${id3}`
-      );
+      const response = await Api.delete(`admin/datasekolah/${id}/siswa/${id2}`);
       data.value.splice(index, 1);
       Toast.success("Success", "Data Berhasil dihapus!");
       return response.data;
@@ -169,68 +166,23 @@ const doPilihKelas = () => {
     <div>
       <span
         class="text-2xl sm:text-3xl leading-none font-bold text-gray-700 shadow-sm"
-        >Catatan Kasus Siswa
-      </span>
+        >Catatan Pengembangandiri Siswa</span
+      >
     </div>
     <div class="md:py-0 py-4 space-x-2 space-y-2"></div>
   </div>
-  <div class="md:pt-6">
-    <div class="md:flex justify-between px-10">
-      <div class="space-x-1 space-y-1 pt-1 md:pt-0">
-        <router-link
-          :to="{
-            name: 'AdminSekolahDetailCatatanKasusSiswaTambah',
-            params: { id, id2 },
-          }"
-        >
-          <button
-            class="btn btn-info hover:shadow-lg shadow text-white hover:text-gray-100 gap-2"
-          >
-            <svg
-              xmlns="http://www.w3.org/2000/svg"
-              class="h-6 w-6"
-              fill="none"
-              viewBox="0 0 24 24"
-              stroke="currentColor"
-            >
-              <path
-                stroke-linecap="round"
-                stroke-linejoin="round"
-                stroke-width="2"
-                d="M4.318 6.318a4.5 4.5 0 000 6.364L12 20.364l7.682-7.682a4.5 4.5 0 00-6.364-6.364L12 7.636l-1.318-1.318a4.5 4.5 0 00-6.364 0z"
-              />
-            </svg>
-            Tambah
-          </button></router-link
-        >
-      </div>
-      <div class="space-x-1 space-y-1 pt-1 md:pt-0">
-        <router-link
-          :to="{
-            name: 'AdminSekolahDetailCatatanKasusSiswa',
-            params: { id: id },
-          }"
-        >
-          <button
-            class="btn hover:shadow-lg shadow text-white hover:text-gray-100 gap-2"
-          >
-            <svg
-              xmlns="http://www.w3.org/2000/svg"
-              class="h-6 w-6"
-              fill="none"
-              viewBox="0 0 24 24"
-              stroke="currentColor"
-            >
-              <path
-                stroke-linecap="round"
-                stroke-linejoin="round"
-                stroke-width="2"
-                d="M4.318 6.318a4.5 4.5 0 000 6.364L12 20.364l7.682-7.682a4.5 4.5 0 00-6.364-6.364L12 7.636l-1.318-1.318a4.5 4.5 0 00-6.364 0z"
-              />
-            </svg>
-            Kembali
-          </button></router-link
-        >
+  <div class="w-full bg-white shadow shadow-md py-4 px-4">
+    <div class="flex justify-center">
+      <v-select
+        class="py-2 px-3 w-72 mx-auto md:mx-0"
+        :options="pilihKelas"
+        v-model="inputCariKelas"
+        v-bind:class="{ disabled: false }"
+      ></v-select>
+      <div class="py-2">
+        <button class="btn btn-sm btn-info p-2" @click="doPilihKelas()">
+          Cari
+        </button>
       </div>
     </div>
   </div>
@@ -255,12 +207,46 @@ const doPilihKelas = () => {
             <template #table-row="props">
               <span v-if="props.column.field == 'actions'">
                 <div
-                  class="text-sm font-medium text-center flex justify-center space-x-0"
+                  class="text-sm font-medium text-center flex justify-center space-x-1"
                 >
-                  <ButtonEdit @click="doEditData(props.row.id, props.index)" />
-                  <ButtonDelete
-                    @click="doDeleteData(props.row.id, props.index)"
-                  />
+                  <button class="btn btn-sm">
+                    <svg
+                      xmlns="http://www.w3.org/2000/svg"
+                      class="h-5 w-5"
+                      fill="none"
+                      viewBox="0 0 24 24"
+                      stroke="currentColor"
+                      stroke-width="2"
+                    >
+                      <path
+                        stroke-linecap="round"
+                        stroke-linejoin="round"
+                        d="M17 17h2a2 2 0 002-2v-4a2 2 0 00-2-2H5a2 2 0 00-2 2v4a2 2 0 002 2h2m2 4h6a2 2 0 002-2v-4a2 2 0 00-2-2H9a2 2 0 00-2 2v4a2 2 0 002 2zm8-12V5a2 2 0 00-2-2H9a2 2 0 00-2 2v4h10z"
+                      />
+                    </svg>
+                  </button>
+                  <router-link
+                    :to="{
+                      name: 'AdminSekolahDetailCatatanPengembanganDiriDetail',
+                      params: { id: id, id2: props.row.id },
+                    }"
+                  >
+                    <button class="btn btn-sm btn-primary">
+                      <svg
+                        xmlns="http://www.w3.org/2000/svg"
+                        class="h-5 w-5"
+                        fill="none"
+                        viewBox="0 0 24 24"
+                        stroke="currentColor"
+                        stroke-width="2"
+                      >
+                        <path
+                          stroke-linecap="round"
+                          stroke-linejoin="round"
+                          d="M12 9v3m0 0v3m0-3h3m-3 0H9m12 0a9 9 0 11-18 0 9 9 0 0118 0z"
+                        />
+                      </svg></button
+                  ></router-link>
                 </div>
               </span>
